+++ pyxel/detectors/detector.py
@@ -19,6 +19,7 @@ from pyxel.data_structure import (
     Persistence,
     Photon,
     Pixel,
+    ProcessedData,
     Scene,
     Signal,
     SimplePersistence,
@@ -48,6 +49,7 @@ class Detector:
         self._pixel: Optional[Pixel] = None
         self._signal: Optional[Signal] = None
         self._image: Optional[Image] = None
+        self._processed_data: Optional[ProcessedData] = None
 
         # This will be the memory of the detector where trapped charges will be saved
         self._memory: Dict = {}
@@ -69,6 +71,7 @@ class Detector:
             and self._pixel == other._pixel
             and self._signal == other._signal
             and self._image == other._image
+            and self._processed_data == other._processed_data
         )
 
     @property
@@ -136,6 +139,14 @@ class Detector:
 
         return self._image
 
+    @property
+    def processed_data(self) -> ProcessedData:
+        """TBW."""
+        if not self._processed_data:
+            raise RuntimeError("'processed_data' not initialized.")
+
+        return self._processed_data
+
     def to_xarray(self) -> "xr.Dataset":
         """Create a new ``Dataset`` from all data containers.
 
@@ -178,6 +189,9 @@ class Detector:
         self._pixel = Pixel(geo=self.geometry)
         self._signal = Signal(geo=self.geometry)
         self._image = Image(geo=self.geometry)
+
+        if self._processed_data is None:
+            self._processed_data = ProcessedData()
 
     def empty(self, empty_all: bool = True) -> None:
         """Empty the data in the detector.
@@ -411,6 +425,7 @@ class Detector:
             "_pixel",
             "_signal",
             "_image",
+            "_processed_data",
             "material",
             "environment",
             "_geometry",