+++ pyxel/detectors/mkid/mkid.py
@@ -131,12 +131,19 @@ class MKID(Detector):
                 "signal": None if self._signal is None else self._signal.array.copy(),
                 "image": None if self._image is None else self._image.array.copy(),
                 "phase": None if self._phase is None else self._phase.array.copy(),
-                "charge": None
-                if self._charge is None
-                else {
-                    "array": self._charge.array.copy(),
-                    "frame": self._charge.frame.copy(),
-                },
+                "processed_data": (
+                    None
+                    if self._processed_data is None
+                    else self._processed_data.data.to_dict()
+                ),
+                "charge": (
+                    None
+                    if self._charge is None
+                    else {
+                        "array": self._charge.array.copy(),
+                        "frame": self._charge.frame.copy(),
+                    }
+                ),
             },
         }
 
@@ -148,6 +155,7 @@ class MKID(Detector):
         """Create a new instance of `MKID` from a `dict`."""
         # TODO: This is a simplistic implementation. Improve this.
         import numpy as np
+        import xarray as xr
 
         from pyxel.data_structure import Scene
         from pyxel.detectors import Environment, MKIDCharacteristics, MKIDGeometry
@@ -184,6 +192,8 @@ class MKID(Detector):
             detector.signal.array = np.asarray(data["signal"])
         if "image" in data:
             detector.image.array = np.asarray(data["image"])
+        if "processed_data" in data:
+            detector.processed_data._data = xr.Dataset.from_dict(data["processed_data"])
         if "charge" in data and data["charge"] is not None:
             charge_dct = data["charge"]
             detector.charge._array = np.asarray(charge_dct["array"])