+++ tests/detectors/test_ccd.py
@@ -149,7 +149,7 @@ def comparison(dct, other_dct):
     assert (
         set(dct["data"])
         == set(other_dct["data"])
-        == {"photon", "scene", "pixel", "signal", "image", "charge"}
+        == {"photon", "scene", "pixel", "signal", "image", "charge", "processed_data"}
     )
     np.testing.assert_equal(dct["data"]["photon"], other_dct["data"]["photon"])
     np.testing.assert_equal(dct["data"]["pixel"], other_dct["data"]["pixel"])
@@ -227,6 +227,12 @@ def comparison(dct, other_dct):
                             dtype=float,
                         ),
                     },
+                    "processed_data": {
+                        "coords": {},
+                        "attrs": {},
+                        "dims": {},
+                        "data_vars": {},
+                    },
                 },
             },
             id="Default parameters",
@@ -297,6 +303,12 @@ def comparison(dct, other_dct):
                             ],
                             dtype=float,
                         ),
+                    },
+                    "processed_data": {
+                        "coords": {},
+                        "attrs": {},
+                        "dims": {},
+                        "data_vars": {},
                     },
                 },
             },
@@ -394,6 +406,7 @@ def test_to_and_from_dict_with_arrays_no_frame(valid_ccd: CCD, klass):
                     dtype=float,
                 ),
             },
+            "processed_data": {"coords": {}, "attrs": {}, "dims": {}, "data_vars": {}},
         },
     }