+++ tests/detectors/test_mkid.py
@@ -150,7 +150,16 @@ def comparison(dct, other_dct):
     assert (
         set(dct["data"])
         == set(other_dct["data"])
-        == {"photon", "scene", "pixel", "signal", "image", "phase", "charge"}
+        == {
+            "photon",
+            "scene",
+            "pixel",
+            "signal",
+            "image",
+            "phase",
+            "charge",
+            "processed_data",
+        }
     )
     np.testing.assert_equal(dct["data"]["photon"], other_dct["data"]["photon"])
     np.testing.assert_equal(dct["data"]["pixel"], other_dct["data"]["pixel"])
@@ -230,6 +239,12 @@ def comparison(dct, other_dct):
                             dtype=float,
                         ),
                     },
+                    "processed_data": {
+                        "coords": {},
+                        "attrs": {},
+                        "dims": {},
+                        "data_vars": {},
+                    },
                 },
             },
             id="Default Parameters",
@@ -300,6 +315,12 @@ def comparison(dct, other_dct):
                             ],
                             dtype=float,
                         ),
+                    },
+                    "processed_data": {
+                        "coords": {},
+                        "attrs": {},
+                        "dims": {},
+                        "data_vars": {},
                     },
                 },
             },